+++ wagtailgeowidget/blocks.py
@@ -1,7 +1,6 @@
-import six
 from django import forms
 from django.utils.functional import cached_property
-from wagtail.core.blocks import CharBlock, FieldBlock
+from wagtail.core.blocks import CharBlock, FieldBlock, IntegerBlock
 
 from wagtailgeowidget.helpers import geosgeometry_str_to_struct
 from wagtailgeowidget.widgets import GeoField
@@ -12,13 +11,28 @@ class GeoAddressBlock(CharBlock):
         classname = "geo-address-block"
 
 
+class GeoZoomBlock(IntegerBlock):
+    class Meta:
+        classname = "geo-zoom-block"
+
+
 class GeoBlock(FieldBlock):
     class Meta:
         icon = "site"
 
-    def __init__(self, address_field=None, required=True, help_text=None, **kwargs):
+    def __init__(
+        self,
+        address_field=None,
+        zoom_field=None,
+        required=True,
+        help_text=None,
+        hide_latlng=False,
+        **kwargs
+    ):
         self.field_options = {}
         self.address_field = address_field
+        self.zoom_field = zoom_field
+        self.hide_latlng = hide_latlng
         super(GeoBlock, self).__init__(**kwargs)
 
     @cached_property
@@ -28,6 +42,8 @@ class GeoBlock(FieldBlock):
                 srid=4326,
                 id_prefix="",
                 address_field=self.address_field,
+                zoom_field=self.zoom_field,
+                hide_latlng=self.hide_latlng,
                 used_in="GeoBlock",
             )
         }
@@ -51,7 +67,7 @@ class GeoBlock(FieldBlock):
         if not value:
             return None
 
-        if value and isinstance(value, six.string_types):
+        if value and isinstance(value, str):
             return value
 
         val = "SRID={};POINT({} {})".format(