+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - New: Make it possible to hide latlng field for GeoBlock (@vladox)
 - Fix: Solve issue with address not working streamfield in Wagtail 1.13+ (@vladox)
+- Fix: Drop six dependency (Martin Sandström)
 
 
 ## 5.2.0 (2022.01.04)