+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## (In development) 5.3.0 (XXXX.XX.XX)
 
-- New: Make it possible to hide latlng field for GeoBlock (@vladox)
+- Enable loading Google Maps API key dynamically (Martin Sandström)
+- Make it possible to hide latlng field for GeoBlock (@vladox)
 - Fix: Solve issue with address not working streamfield in Wagtail 1.13+ (@vladox)
 - Fix: Drop six dependency (Martin Sandström)