+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
 
+## (In development) 5.3.0 (2020.01.05)
+
+- Add persistant and user editable zoom for map widget (Martin Sandström)
+- Enable loading Google Maps API key dynamically (Martin Sandström)
+- Make it possible to hide latlng field for GeoBlock (@vladox)
+- Fix: Solve issue with address not working streamfield in Wagtail 1.13+ (@vladox)
+- Fix: Drop six dependency (Martin Sandström)
+
+
 ## 5.2.0 (2022.01.04)
 
 - Drop support for Python 3.6